@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { SpeechNotification } from '../model/speech-notification';
 import { SpeechError } from '../model/speech-error';
@@ -9,6 +9,7 @@ const { webkitSpeechRecognition }: AppWindow = <AppWindow>window;
 
 @Injectable()
 export class SpeechRecognizerService {
+
   recognition: any;
   startTimestamp;
   ignoreOnEnd: boolean;
@@ -65,18 +66,16 @@ export class SpeechRecognizerService {
   onResult(): Observable<SpeechNotification> {
     return new Observable(observer => {
       this.recognition.onresult = (event) => {
-        let interimTranscript = '',
-          interimSpan,
-          finalSpan,
-          finalTranscript = '';
-        for (var i = event.resultIndex; i < event.results.length; ++i) {
+        let interimTranscript = '';
+        let finalTranscript = '';
+
+        for (let i = event.resultIndex; i < event.results.length; ++i) {
           if (event.results[i].isFinal) {
             finalTranscript += event.results[i][0].transcript;
           } else {
             interimTranscript += event.results[i][0].transcript;
           }
         }
-        interimSpan = this.linebreak(interimTranscript);
 
         observer.next({
           info: 'final_transcript',
@@ -94,15 +93,15 @@ export class SpeechRecognizerService {
     return new Observable(observer => {
       this.recognition.onerror = (event) => {
         let result: SpeechError;
-        if (event.error == 'no-speech') {
+        if (event.error === 'no-speech') {
           result = SpeechError.NO_SPEECH;
           this.ignoreOnEnd = true;
         }
-        if (event.error == 'audio-capture') {
+        if (event.error === 'audio-capture') {
           result = SpeechError.NO_MICROPHONE;
           this.ignoreOnEnd = true;
         }
-        if (event.error == 'not-allowed') {
+        if (event.error === 'not-allowed') {
           if (event.timeStamp - this.startTimestamp < 100) {
             result = SpeechError.BLOCKED;
         } else {
@@ -120,16 +119,5 @@ export class SpeechRecognizerService {
 
   stop() {
     this.recognition.stop();
-  }
-
-  private capitalize(s: string) {
-    const firstChar = /\S/;
-    return s.replace(firstChar, function (m) { return m.toUpperCase(); });
-  }
-
-  private linebreak(s) {
-    const twoLine = /\n\n/g;
-    const oneLine = /\n/g;
-    return s.replace(twoLine, '<p></p>').replace(oneLine, '<br>');
   }
 }
