@@ -1,17 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import {
-  MatButtonModule,
-  MatCardModule,
-  MatInputModule,
-  MatFormFieldModule,
-  MatIconModule,
-  MatToolbarModule,
-  MatSelectModule,
-  MatListModule,
-  MatSidenavModule
-} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatSelectModule } from '@angular/material/select';
+import { MatListModule } from '@angular/material/list';
+import { MatSidenavModule } from '@angular/material/sidenav';
 
 @NgModule({
   imports: [
@@ -36,8 +34,6 @@ import {
     MatSelectModule,
     MatListModule,
     MatSidenavModule
-  ],
-  providers: [],
-  declarations: []
+  ]
 })
 export class MaterialModule { }
