@@ -1,6 +1,6 @@
 export enum SpeechError {
-    NO_SPEECH,
-    NO_MICROPHONE,
-    NOT_ALLOWED,
-    BLOCKED
+  NO_SPEECH,
+  NO_MICROPHONE,
+  NOT_ALLOWED,
+  BLOCKED
 }
