@@ -1,6 +1,7 @@
 import { SpeechError } from './speech-error';
+
 export interface SpeechNotification {
     info?: string;
     error?: SpeechError;
-    content?:any;
+    content?: any;
 }
