@@ -1,6 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { WebSpeechComponent } from './web-speech.component';
+import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { SpeechRecognizerService } from './shared/services/speech-recognizer.service';
+import { SpeechSynthesizerService } from './shared/services/speech-synthesizer.service';
 
 describe('WebSpeechComponent', () => {
   let component: WebSpeechComponent;
@@ -8,7 +11,16 @@ describe('WebSpeechComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ WebSpeechComponent ]
+      declarations: [
+        WebSpeechComponent
+      ],
+      schemas: [
+        CUSTOM_ELEMENTS_SCHEMA
+      ],
+      providers: [
+        SpeechRecognizerService,
+        SpeechSynthesizerService
+      ]
     })
     .compileComponents();
   }));
